# docs/swagger.yaml
definitions:
...
  project.InjectOperation:
    properties:
      archived:
        type: boolean
      customFilenames:
        items:
          type: string
        type: array
      customPayloads:
        items:
          type: string
        type: array
      doNotRecord:
        type: boolean
      error:
        type: string
      fuzzDB:
        items:
          type: string
        type: array
      guid:
        type: string
      host:
        type: string
      injectDescription:
        type: string
      iterateFrom:
        type: integer
      iterateTo:
        type: integer
      objectType:
        type: string
      percentCompleted:
        description: Parts of the object which cannot be set by JSON
        type: integer
      request:
        items:
          $ref: '#/definitions/project.InjectOperationRequestPart'
        type: array
      requestsMadeCount:
        type: integer
      ssl:
        type: boolean
      title:
        type: string
      totalRequestCount:
        type: integer
      url:
        type: string
    type: object
  project.InjectOperationRequestPart:
    properties:
      inject:
        type: boolean
      requestPart:
        type: string
    type: object
...
host: localhost
info:
  contact:
    email: support@forensant.com
    name: API Support
    url: https://forensant.com/support
  description: This provides the common functions which are relied upon by the Proximity
    frontends.
  license:
    name: Commercial
    url: https://proximity.forensant.com/terms
  termsOfService: https://forensant.com/terms/
  title: Proximity Core
  version: "1.0"
paths:
...
  /inject_operations:
    get:
      description: gets a list of all injection operations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/project.InjectOperation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get All Inject Operations
      tags:
      - Injection Operations
  /inject_operations/{guid}/archive:
    patch:
      description: updates the the archived status of an inject operation
      parameters:
      - description: inject operation guid
        in: path
        name: guid
        required: true
        type: string
      - description: archive status to set
        in: formData
        name: archive
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Archive Inject Operation
      tags:
      - Injection Operations
  /inject_operations/{guid}/title:
    patch:
      description: updates the title of an inject operation
      parameters:
      - description: inject operation guid
        in: path
        name: guid
        required: true
        type: string
      - description: title to set
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Set Inject Operation Title
      tags:
      - Injection Operations
  /inject_operations/{path}:
    get:
      description: gets a single inject operation
      parameters:
      - description: The GUID of the request to fetch
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.InjectOperation'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get Inject Operation
      tags:
      - Injection Operations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"